+++ sdmx-dl-provider-util/src/main/java/internal/sdmxdl/util/ext/EcbDialect.java
@@ -16,7 +16,6 @@
  */
 package internal.sdmxdl.util.ext;
 
-import nbbrd.io.text.Parser;
 import nbbrd.service.ServiceProvider;
 import org.checkerframework.checker.nullness.qual.NonNull;
 import sdmxdl.DataStructure;
@@ -26,12 +25,10 @@ import sdmxdl.ext.ObsParser;
 import sdmxdl.ext.spi.SdmxDialect;
 import sdmxdl.util.parser.DefaultObsParser;
 import sdmxdl.util.parser.FreqFactory;
-import sdmxdl.util.parser.TimeFormatParsers;
 
 import java.util.Objects;
 
 import static sdmxdl.Frequency.*;
-import static sdmxdl.util.parser.TimeFormatParsers.FIRST_DAY_OF_YEAR;
 
 /**
  * @author Philippe Charles
@@ -56,11 +53,10 @@ public final class EcbDialect implements SdmxDialect {
 
     private static ObsParser getObsParser(DataStructure dsd) {
         Objects.requireNonNull(dsd);
-        return new DefaultObsParser(
-                getFreqFactory(dsd),
-                freq -> TimeFormatParsers.getObservationalTimePeriod(FIRST_DAY_OF_YEAR),
-                Parser.onDouble()
-        );
+        return DefaultObsParser
+                .builder()
+                .freqFactory(getFreqFactory(dsd))
+                .build();
     }
 
     private static FreqFactory getFreqFactory(DataStructure dsd) {